Always send out mails with the current date, not the date of the commit.
post-receive-diff can end up sending emails for commits that happen
distant in the past, eg. when they've been sitting on a developer
machine for some time without pushing, or when merging a long-lived
branch.

Previously, the original Date header of the commit was used in outgoing
mail, which had the undesirable side effect of breaking expectations of
MUAs of sorting by date being appropriate in order to show new mail
together at the top (or the bottom).

This commit changes post-receive-diff to always use the current time for
the Date header of outgoing mail. The original timestamp of the commit
is preserved in the Original-Date header.