tag-pending: use committer name instead of author name.
Setting a pending tag is an operation that the maintainer does. It is
the maintainer who pushes a change to the repository, but the patch may
actually be authored by somebody else. In that case, I think it'd be
indeed undesirable for the "tag + pending" message to come from a person
who is not a maintainer of the package.

Additionally, the author information is included in the body of the
sent message, together with the message of the commit itself.

Suggested-by: Raphael Hertzog <hertzog@debian.org>